backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
openai==1.35.3
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
pydantic==2.7.3
pydantic-settings==2.3.1
python-multipart==0.0.9
python-dotenv==1.0.1